Fix FITS configuration setting
The config.fits_path setting was being set correctly but,
unfortunately, was not being enabled because it was still commented out
in config/initializers/sufia.rb.  This fix makes sure the
config.fits_path setting is uncommented as well as being set to point
to the correct FITS application installed earlier.