Support cloning private application code repository
Add support for cloning a public or private repository containing the
Sufia application code.  The HYDRA_HEAD_GIT_REPO_DEPLOY_KEY is used to
specify a SSH private key to be used when cloning a private repository.

This change has also enabled SSH Agent forwarding in the Vagrantfile to
enable developers who have SSH Agent access to a private repository
from their local machine to be able to clone it as part of the install
without having to use an explicit deployment key.