Fix install so "bootstrap.sh aws" works for ETDplus
With the introduction of support for deployment from private Git
repositories, the size of the install scripts and files/ material
packed into the aws_bootstrapper.sh script exceeded the 16KB allowed
for user-data.  (This was particularly true when a SSH deploy key was
actually being used with the bootstrap.)  This caused the creation of
AWS instances to fail.

As a fix for this, the bootstrap_aws() function of bootstrap.sh was
changed so that shell comments are removed from install*.sh and
config{,_aws}.sh files before being packed into aws_bootstrapper.sh.
The bootstrap_server.sh script is not stripped because it contains a
"here document" that includes structured comments that are not safe to
strip nor are straightforward to avoid stripping.  Shell comments are
defined to be lines beginning with "#" (optionally preceded by
whitespace), i.e., whole line comments, not comments appended to the
end of shell code on a line.

To make the comment removal more effective, the comments in the
config{,_aws}.sh files were reformatted so they would be capable of
being stripped.  This was also taken as an opportunity to expand the
comments in those files to describe more of their contents.

TMPDIR was changed in bootstrap_sh() because it has special meaning to
mktemp and shouldn't be used in the context in which it was being used.
 It was changed to AWSDIR to avoid conflict.