Don't install redundant gems (i.e., Rails)
Currently, we explicitly install the Rails gem but then, after fetching
the application code, we execute "bundle install" to install the
required gems used by the application.  (These gems are specified in
Gemfile/Gemfile.lock.)  The earlier installation of the Rails gem may
not install the same version as specified by the application in
Gemfile.lock.  The latter version is the one that will be used, so it
is redundant to install the Rails gem explicitly, as done earlier.

Another disadvantage of installing the Rails gem explicitly is that it
had the side effect of installing the Bundler gem.  The Bundler gem is
a vital prerequisite for doing "bundle install" and so it should be
installed explicitly.  This change does that instead of installing the
Rails gem explicitly.