Install FITS to /opt/fits
Currently, FITS is installed to $FITS_DIR, which is usually under the
$INSTALL_DIR.  The path to the "fits.sh" script is set in the Sufia
initialiser, config/initializers/sufia.rb.  This is done during install
by the install scripts, which change the value in the code to point to
the installed "fits.sh" script.

This change is so the latter fix-up is not necessary.  Instead of
installing to $FITS_DIR, we install FITS to a fixed location:
/opt/fits.  (We actually install to /opt/$FITS_PACKAGE and symlink
/opt/fits to there.  That would allow us in future potentially to have
several installed versions of FITS and point to the one we are using
via the symlink, as Solr does.)