Create default settings in secrets.yml
This changes the install process so that configuration settings are
supplied in "secrets.yml" and read from there by the application.  This
means the "fix-ups" formerly required that altered application code or
config files are no longer needed.  This makes updating the application
code easier, as merge conflicts are unlikely.

The "secrets.yml.sample" file contains a default sample "secrets.yml"
file that can be put into the "files/" directory (after renaming to
"secrets.yml").  If no "secrets.yml" file is present in "files/" during
bootstrapping, a default one is created.  The default "secrets.yml"
generated also contains the important "secret_key_base" key required by
Rails for the application to run.  It is randomly generated.  (It is
the responsibility of the user to supply this if using a custom
"secrets.yml" file.)