Support local/remote PostgreSQL databases
This change introduces a DB_IS_REMOTE configuration option to designate
whether the PostgreSQL database server is running on a remote system or
should be installed to run locally.

If DB_IS_REMOTE="YES" then we skip installing the PostgreSQL server as
part of the installation process.  Instead, we install only the
PostgreSQL client applications and then use them to create the
application database and user on the remote server.

Otherwise, if DB_IS_REMOTE is not "YES" then it is assumed that a
PostgreSQL server will be installed locally.

When installing with DB_IS_REMOTE="YES", it is incumbent upon the
person installing to supply an appropriately configured
files/secrets.yml file.  In particular, the "host:" setting in the
"database:" section of files/secrets.yml must be set accordingly for
the installation to succeed.  (Otherwise, the DB load task will
fail during installation.)  This is in addition to specifying the
appropriate DB_* configuration options in config.sh, etc.