Provision AWS using Vagrant
Previously, a bootstrap.sh script was responsible for kickstarting the
installation, getting things ready so that bootstrap_server.sh could
handle the actual server and application installation.  For AWS this
meant bundling up the installation scripts into the "user data"
supplied when starting up the EC2 instance.  The "user data" is limited
in size, and the current installation scripts tend to exceed this size
regularly now.

Based upon the VTUL/sufia-install approach, this change is to have
Vagrant handle all the provisioning.  This means eliminating
bootstrap.sh and handling the initial server setup via the Vagrantfile.

Now, instead of using bootstrap.sh the user brings up the server using
"vagrant up" and uses the "--provider" option to determine whether the
server is to be brought up locally via VirtualBox or as an EC2 instance
in AWS.  (Note that the vagrant-aws plugin is required to be installed
locally to use the "--provider aws" option.)

Other than that, the rest of the installation is the same as before.
Now, like the local VM install, the AWS install leverages the /vagrant
shared folder mechanism to locate the scripts on the server being
brought up.  Also, Vagrant handles bringing up the AWS server and
copying the scripts to it, meaning there is no longer any need for the
AWS bootstrapper and much of the work done by bootstrap.sh.